Consolidate MimicExplainer serialization tests
